Correct event lines routing from HWPE to cluster
2 QuestaSim warnings revealed that event lines from HWPE accelerators to
the cluster had a width mismatch (error code: 3015).

The errors were due to the following bugs:
- Events were truncated in the cluster interface. They should be instead
passed full-width to the higher hierarchy (LIC accelerator region);
- Events (and busy signals) were processed with packed arrays. Instead,
the expected behaviour was better matching to unpacked arrays.
In fact, events are grouped from HWPEs are grouped and passed to an OR, where the output is
the event in the cluster domain (a single accelerator event at cluster-level).